pyproject.toml
[tool.poetry]
name = "data"
version = "0.1.0"
description = "grab data from cloud services"
authors = ["Raymond Chang"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gdown = "^4.7.1"
rclone = "^0.4.4"
oauth2client = "^4.1.3"
google-api-python-client = "^2.103.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
pytest = "^7.4.2"
pytest-mock = "^3.11.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
